% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_reach.R
\name{auth_reach}
\alias{auth_reach}
\title{auth_reach}
\usage{
auth_reach(
  user = Sys.getenv("REACH_API_USER"),
  pw = Sys.getenv("REACH_API_PW")
)
}
\arguments{
\item{user}{Username for Reach API}

\item{pw}{Password for Reach API}
}
\value{
An access token for use in other functions
}
\description{
Function to authenticate with the Reach API and create access token for other functions
}
