% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ef_geom.R
\name{check_ef_geom}
\alias{check_ef_geom}
\title{Check if sf geometry is suitable for importing into Emlid Flow}
\usage{
check_ef_geom(x)
}
\arguments{
\item{x}{An sf object.}
}
\value{
logical. TRUE indicates that current geometry type is compatible for import.
}
\description{
Check if sf geometry is suitable for importing into Emlid Flow
}
\details{
Emlid Flow currently supports only (multi)point geometries.
}
\seealso{
\code{\link[=as_ef_geom]{as_ef_geom()}}
}
