% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rtk_range.R
\name{check_rtk_range}
\alias{check_rtk_range}
\title{Check that points or area of interest are within range of proposed RTK base.}
\usage{
check_rtk_range(
  rover,
  base,
  local_crs = getOption("local_crs"),
  range = getOption("lora_range", 8000),
  plot = TRUE
)
}
\arguments{
\item{rover}{sf object. Point(s) or polygon(s) containing the area of interest where the rover will be used.}

\item{base}{sf object. A point containing the proposed coordinates of the RTK base.}

\item{local_crs}{String. A projected coordinate reference system, either an EPSG code, WKT, or proj4string.}

\item{range}{Numeric. The nominal range of the LoRa signal in meters, defaults to 8000 m (=8 km).}

\item{plot}{Logical. Whether or not}
}
\value{
Logical, also a message reporting the result and (optionally) a ggplot.
}
\description{
Check that points or area of interest are within range of proposed RTK base.
}
