% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnld_flow360.R
\name{dnld_flow360}
\alias{dnld_flow360}
\title{Download Emlid Flow360 projects}
\usage{
dnld_flow360(
  dir = "~/Downloads",
  overwrite = FALSE,
  usr = Sys.getenv("EMLID_USER"),
  pwd = Sys.getenv("EMLID_PASSWORD"),
  mult = 1,
  scroll_step = 250,
  max_scrolls = 100
)
}
\arguments{
\item{dir}{Directory where the downloaded csv files should be saved.}

\item{overwrite}{Logical, whether or not to overwrite existing files with the same name. NOT YET IMPLEMENTED.}

\item{usr}{Emlid credential username.}

\item{pwd}{Emlid credential password.}

\item{mult}{Multiplier of default wait times.}
}
\value{
(Invisible) Files will be downloaded to the specified directory.
}
\description{
Download Emlid Flow360 projects
}
\examples{
dnld_flow360(dir = "~/Downloads")

}
