% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_external_id_types.R
\name{get_external_id_types}
\alias{get_external_id_types}
\title{get_external_id_types}
\usage{
get_external_id_types()
}
\value{
A dataframe of external ID types
}
\description{
Function to retrieve external ID types for Reach API
}
