% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{get_tags}
\usage{
get_tags()
}
\value{
A dataframe of all tags in the Reach campaign for which you are authenticated
}
\description{
Function to retrieve tags from Reach API
}
