% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tag.R
\name{import_tag}
\alias{import_tag}
\title{import_tag}
\usage{
import_tag(tag_id, file_url)
}
\arguments{
\item{tag_id}{ID of the tag you want to apply}

\item{file_url}{URL of the hosted CSV file}
}
\value{
A message indicating the import has started
}
\description{
Function to apply tags from a hosted CSV file into voters in Reach
}
