% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tag_status.R
\name{import_tag_status}
\alias{import_tag_status}
\title{import_tag_status}
\usage{
import_tag_status(status_url)
}
\arguments{
\item{status_url}{URL for tag import job status}
}
\value{
A message indicating the import job's status
}
\description{
Function to retrieve tag import job status from Reach API
}
