% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_vrs.R
\name{mk_vrs}
\alias{mk_vrs}
\title{Convert LLH log file to inputs for downloading CORS/VRS corrections}
\usage{
mk_vrs(llh, rnd = 0, ...)
}
\arguments{
\item{llh}{File path to .LLH log file}

\item{rnd}{Number of minutes to round to on both ends of the logging duration (see Details)}

\item{...}{Additional arguments passed to utils::read.table()}
}
\value{
a list containing the average logged position and
}
\description{
Convert LLH log file to inputs for downloading CORS/VRS corrections
}
\details{
The 'rnd' argument helps pad the logging interval to ensure that the VRS data include
the entire logging duration. A value of 0 will not pad either the start or end times,
resulting in an exact start time and duration of the LLH log. Positive values will
ceiling round the start time and floor round the stop time to the number of minutes
specified by the user. For example, an interval of 10:28:53-13:56:23 will round as
follows:
rnd = 0: 10:28:53-13:56:23, a duration of 03:28:30
rnd = 1: 10:28:00-13:57:00, a duration of 03:29:00
rnd = 15: 10:15:00-14:00:00, a duration of 04:15:00
For additional information on rounding behaviors, see ?lubridate::round_date.
}
\examples{
\dontrun{
mk_vrs(llh = "path/to/file.llh")
}

}
