% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachr-package.R
\docType{package}
\name{reachr-package}
\alias{reachr}
\alias{reachr-package}
\title{reachr: Tools for Interacting with Survey Points Collected by Emlid GNSS Products}
\description{
Utility functions for interacting with data produced by Emlid Reach RS2(+) and Emlid Studio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DICELab-NCSU/reachr}
  \item Report bugs at \url{https://github.com/DICELab-NCSU/reachr/issues}
}

}
\author{
\strong{Maintainer}: William K. Petry \email{wpetry@ncsu.edu} (\href{https://orcid.org/0000-0002-5230-5987}{ORCID})

}
\keyword{internal}
