% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emlid.R
\name{read_emlid}
\alias{read_emlid}
\title{Read files produced by Emlid software as simple features}
\usage{
read_emlid(
  file,
  format = c("guess", "shp", "csv", "csv_penzd", "dxf", "pos"),
  crs = NULL,
  drop_z = FALSE,
  ...
)
}
\arguments{
\item{file}{string. Path to file containing spatial data.}

\item{format}{string. The file format of the data. Defaults to guessing based on the file
extension.}

\item{crs}{string. The coordinate reference system for the spatial data, supplied as an
EPSG code, WKT, or proj4string.}

\item{drop_z}{logical, whether or not to drop the z/vertical dimension from points.
Defaults to FALSE.}

\item{...}{passed to function that reads in the datafile}
}
\value{
object of class sf
}
\description{
Read files produced by Emlid software as simple features
}
