% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinex_summary.R
\name{rinex_summary}
\alias{rinex_summary}
\title{Summarize Rinex observation file}
\usage{
rinex_summary(rinex)
}
\arguments{
\item{rinex}{Path to Rinex observation file (.YYo where YY are the last two digits of the observation year).}
}
\value{
A list containing the Rinex version, number of satellites in used from each constellation, and the observation start/end times.
}
\description{
Summarize Rinex observation file
}
\examples{
\dontrun{
rinex_summary(rinex = "path/to/file.25o")
}
}
