% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_base.R
\name{update_base}
\alias{update_base}
\title{Update coordinates to use a new base position}
\usage{
update_base(points, newbase, base_row = NULL, destfile = NULL)
}
\arguments{
\item{points}{path to a file produced using the Emlid Flow csv export tool}

\item{newbase}{path to a file containing the updated base coordinates}

\item{base_row}{(optional) row containing the updated base coordinates if newbase contains
multiple points}

\item{destfile}{(optional) path to write a csv file containing corrected points}
}
\value{
a tibble following the formatting conventions of Emlid Flow csv export
}
\description{
Update coordinates to use a new base position
}
