% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_emlid_csv.R
\name{write_emlid_csv}
\alias{write_emlid_csv}
\title{Write simple features spatial data to an Emlid Flow-compatible CSV file}
\usage{
write_emlid_csv(x, file)
}
\arguments{
\item{x}{An sf object.}

\item{file}{string. Path to file containing spatial data.}
}
\value{
Invisible
}
\description{
Write simple features spatial data to an Emlid Flow-compatible CSV file
}
