% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilaterate.R
\name{multilaterate}
\alias{multilaterate}
\title{Determine target point coordinates from distances to known points}
\usage{
multilaterate(distances, ref_points, crs = NULL)
}
\arguments{
\item{distances}{A named vector containing the distances from at least 3 ref_points to the target point.}

\item{ref_points}{An sf object containing the reference point geometry and a 'Name' column.}

\item{crs}{A projected coordinate reference system to use in the distance calculations if \code{ref_points} does not already have a projected CRS.}
}
\value{
A list, with xy coordinates in \code{coords} and measures of coordinate and positional uncertainty in \code{se}.
}
\description{
Determine target point coordinates from distances to known points
}
\details{
This function performs 2D multilateration via optimization. The optimization will fail or give nonsensical results if the units of \code{distances} and \code{crs} are not identical. It is recommended that both use meters. The standard errors for the x coordinate, y coordinate, and xy positional uncertainty are calculated from the Hessian of the objective function.
}
\examples{
d <- c(A = 7.1, B = 7.2, C = 7.0, D = 7.1)
ref <- sf::st_as_sf(data.frame(
 Name = c("A", "B", "C", "D"),
 x = c(0, 10, 0, 10),
  y = c(0, 0, 10, 10)
), coords = c("x", "y"), crs = 32613)
multilaterate(distances = d, ref_points = ref)
}
